+++ src/components/account/AccountPlan.tsx
@@ -33,7 +33,7 @@ function AccountPlan(props: AccountPlanProps) {
           props.active
             ? 'border-emerald-400 bg-emerald-50'
             : 'border-rose-400 bg-rose-50'
-        }  group my-4 w-full transform cursor-pointer rounded-lg px-4 py-2
+        }  group my-4 w-full transform cursor-default rounded-lg px-4 py-2
             text-left text-black transition duration-300 ease-in-out hover:-translate-y-1
             hover:shadow-md active:opacity-50 dark:bg-gray-800`}
         onClick={() => {