+++ src/components/schedule/ScheduleClass.tsx
@@ -68,7 +68,7 @@ function ScheduleClass(props: ScheduleClassProps) {
     <div
       className={`absolute z-10 rounded-lg bg-opacity-60
                 bg-${color}-100 border-2 border-l-4 dark:bg-gray-800 border-${color}-400 group
-                cursor-pointer overflow-visible transition duration-300 ease-in-out ${
+                cursor-default overflow-visible transition duration-300 ease-in-out ${
                   interactions?.hoverSection.get === section.section_id
                     ? '-translate-y-2 shadow-lg'
                     : ''