+++ CHANGELOG.md
@@ -11,11 +11,15 @@ _2023-01-XX_
 - **Toolbar!**
   - Account stuff can be accessed from the toolbar (sign in, sign out, account settings, etc.).
   - Buttons from the old utility bar have been moved to the toolbar.
+  - Some buttons from the taskbar have been moved to the toolbar.
   - The active plan's or schedule's name is now displayed in the toolbar instead of the taskbar.
 - The account plans and schedules tab has been redesigned.
 - You can now press escape or click outside of the side card to close it.
 - You can now press enter to submit pop-up menus that contain a single text input.
 - The built-in feedback submission feature has been replaced with access to the knowledge base and external feedback submission system.
+- "My List" has been renamed to "Bookmarks".
+- The taskbar now has less buttons and is more visible.
+- Clickable elements now use the main cursor by default.
 
 ### 2.1.3